+++ src/main/java/ca/corefacility/bioinformatics/irida/config/services/IridaApiServicesConfig.java
@@ -60,6 +60,8 @@ import ca.corefacility.bioinformatics.irida.processing.impl.DefaultFileProcessin
 import ca.corefacility.bioinformatics.irida.processing.impl.FastqcFileProcessor;
 import ca.corefacility.bioinformatics.irida.processing.impl.GzipFileProcessor;
 import ca.corefacility.bioinformatics.irida.repositories.analysis.submission.AnalysisSubmissionRepository;
+import ca.corefacility.bioinformatics.irida.repositories.joins.sample.SampleSequencingObjectJoinRepository;
+import ca.corefacility.bioinformatics.irida.repositories.sample.QCEntryRepository;
 import ca.corefacility.bioinformatics.irida.repositories.sequencefile.SequencingObjectRepository;
 import ca.corefacility.bioinformatics.irida.service.AnalysisSubmissionCleanupService;
 import ca.corefacility.bioinformatics.irida.service.TaxonomyService;
@@ -149,20 +151,21 @@ public class IridaApiServicesConfig {
 
 	@Bean
 	public FileProcessingChain fileProcessorChain(SequencingObjectRepository sequencingObjectRepository,
-			GzipFileProcessor gzipFileProcessor, FastqcFileProcessor fastQcFileProcessor,
-			AssemblyFileProcessor assemblyFileProcessor, ChecksumFileProcessor checksumProcessor) {
-		
+			QCEntryRepository qcRepository, GzipFileProcessor gzipFileProcessor,
+			FastqcFileProcessor fastQcFileProcessor, AssemblyFileProcessor assemblyFileProcessor,
+			ChecksumFileProcessor checksumProcessor) {
+
 		gzipFileProcessor.setRemoveCompressedFiles(removeCompressedFiles);
 
-		final List<FileProcessor> fileProcessors = Lists.newArrayList(checksumProcessor, gzipFileProcessor, fastQcFileProcessor,
-				assemblyFileProcessor);
+		final List<FileProcessor> fileProcessors = Lists.newArrayList(checksumProcessor, gzipFileProcessor,
+				fastQcFileProcessor, assemblyFileProcessor);
 
 		if (!decompressFiles) {
 			logger.info("File decompression is disabled [file.processing.decompress=false]");
 			fileProcessors.remove(gzipFileProcessor);
 		}
 
-		return new DefaultFileProcessingChain(sequencingObjectRepository, fileProcessors);
+		return new DefaultFileProcessingChain(sequencingObjectRepository, qcRepository, fileProcessors);
 	}
 	
 	@Bean(name = "fileProcessingChainExecutor")