+++ src/main/java/ca/corefacility/bioinformatics/irida/model/sample/FileProcessorErrorQCEntry.java
@@ -1,0 +1,29 @@
+package ca.corefacility.bioinformatics.irida.model.sample;
+
+import javax.persistence.Entity;
+
+import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
+import ca.corefacility.bioinformatics.irida.processing.FileProcessor;
+
+/**
+ * {@link QCEntry} for a failed {@link FileProcessor}
+ * 
+ * @see FileProcessor
+ */
+@Entity
+public class FileProcessorErrorQCEntry extends QCEntry {
+
+	public FileProcessorErrorQCEntry() {
+		super();
+	}
+
+	public FileProcessorErrorQCEntry(SequencingObject sequencingObject) {
+		super(sequencingObject);
+	}
+
+	@Override
+	public QCEntryType getType() {
+		return QCEntryType.PROCESSING;
+	}
+
+}